# tests/api/openapi_v1_snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The consistent error envelope returned for handled failures (AAEO-9).",
        "properties": {
          "code": {
            "default": "error",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "x-framework-generated": "validation-error-envelope"
      },
...
      "RecommendationItem": {
        "description": "A dashboard-facing advisory output, extracted from a run's structured output.",
        "properties": {
          "confidence": {
            "default": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rationale": {
            "default": "",
            "title": "Rationale",
            "type": "string"
          },
          "recommendation_id": {
            "title": "Recommendation Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus",
            "default": "PROPOSED"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "workspace_id",
          "subject_id",
          "kind",
          "title"
        ],
        "title": "RecommendationItem",
        "type": "object"
      },
      "RecommendationListResponse": {
        "description": "A paginated list of recommendation items.",
        "properties": {
          "items": {
            "items": {},
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "title": "RecommendationListResponse",
        "type": "object"
      },
      "RecommendationStatus": {
        "description": "Lifecycle state of an advisory recommendation extracted from a run.",
        "enum": [
          "PROPOSED",
          "ACCEPTED",
          "DISMISSED",
          "SCHEDULED"
        ],
        "title": "RecommendationStatus",
        "type": "string"
      },
...
      "UpdateRecommendationRequest": {
        "description": "A status/notes transition payload for one recommendation.",
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateRecommendationRequest",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/recommendations": {
      "get": {
        "description": "List recommendations (created_at desc), paginated into an envelope (AAEO-8).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal's list is pinned to its\nOWN subject via :func:`narrow_subject` (a ``tenant_admin`` and every non-subject-scoped\n/ offline / ``all_tenants`` caller keep the caller-supplied filter), so a subject-scoped\ntenant can never enumerate another data subject's subject-attributed recommendations\nwithin a shared workspace \u2014 mirroring the runs router. No-op for the zero-config path.",
        "operationId": "list_recommendations_v1_recommendations_get",
        "parameters": [
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "in": "query",
            "name": "subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecommendationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/recommendations/{recommendation_id}": {
      "patch": {
        "description": "Transition a recommendation's status (404 when unknown/out-of-workspace).\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal that targets a\nrecommendation attributed to ANOTHER data subject within its own tenant gets a clean\n404 (never a mutation, never an existence-confirming 403) \u2014 the workspace scope is\nchecked first, then the recommendation's ``subject_id`` is gated via\n:func:`authorize_object`. No-op for offline / ``all_tenants`` / ``tenant_admin``.",
        "operationId": "update_recommendation_v1_recommendations__recommendation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/recommendations/{recommendation_id}/provenance": {
      "get": {
        "description": "Trace a recommendation back to its run, persona, prompt, and cited evidence.\n\nReturns the typed lineage graph as JSON, or Graphviz DOT with ``?format=dot``.\n404 when the recommendation is unknown / out-of-workspace or was never graphed.\n\nObject-level (BOLA, WS-bola): a ``subject_scoped`` principal that requests the\nprovenance subgraph (persona/prompt/cited-evidence) of a recommendation attributed to\nANOTHER data subject within its own tenant gets a clean 404 \u2014 the recommendation's\n``subject_id`` is resolved and gated via :func:`authorize_object` before any lineage is\ntraced. No-op for offline / ``all_tenants`` / ``tenant_admin`` callers.",
        "operationId": "get_recommendation_provenance_v1_recommendations__recommendation_id__provenance_get",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 64,
              "minimum": 0,
              "title": "Max Depth",
              "type": "integer"
            }
          },
          {
            "description": "comma-separated relation allow-list (e.g. derived_from,cites)",
            "in": "query",
            "name": "relations",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma-separated relation allow-list (e.g. derived_from,cites)",
              "title": "Relations"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "pattern": "^(json|dot)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Recommendation Provenance V1 Recommendations  Recommendation Id  Provenance Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recommendation Provenance",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}